requirements.txt
asgiref==3.5.2
certifi==2022.9.24
coreapi==2.3.3
Django==4.1.2
django-cors-headers==3.13.0
django-filter==23.1
djangorestframework==3.14.0
drf-yasg==1.21.4
pillow==9.3.0
pip==23.0.1
PyJWT==2.6.0
pytz==2022.5
six==1.16.0
